reset script checks

+++ reset_pli.sh
@@ -76,17 +76,37 @@ sudo apt autoremove -y
 sudo rm -rf /usr/lib/postgresql/ && sudo rm -rf /var/lib/postgresql/ && sudo rm -rf /var/log/postgresql/ && sudo rm -rf /etc/postgresql/ && sudo rm -rf /etc/postgresql-common/
 
 # Remove the POSTGRES install system account & group
-sudo userdel -r postgres && sudo groupdel postgres
+#sudo userdel -r postgres && sudo groupdel postgres
+
+if [ $(getent passwd postgres) ]; then
+  sudo userdel -r postgres
+fi
+
+if [ $(getent group postgres) ]; then
+  sudo groupdel postgres
+fi
+
 
 # Remove the group for local backups
-sudo groupdel nodebackup
+if [ $(getent group nodebackup) ]; then
+  sudo groupdel nodebackup
+fi
+
 
 # Remove all plugin, nodejs linked folders for current user & root
 cd ~/; sudo sh -c "rm -rf .cache/ && rm -rf .nvm && rm -rf .npm && rm -rf .plugin && rm -rf Plugin && rm -rf .pm2 && rm -rf work && rm -rf go && rm -rf .yarn*"
-rm ~/.tmp_profile
+
+
+if [ -e ~/.tmp_profile ]; then
+    rm -f ~/.tmp_profile
+fi
 
 # Remove logrotate file
-sudo sh -c 'rm -f /etc/logrotate.d/plugin-logs'
+if [ -e /etc/logrotate.d/plugin-logs ]; then
+    sudo sh -c 'rm -f /etc/logrotate.d/plugin-logs'
+fi
+
+
